requirements.txt
setuptools~=60.5.0
click~=8.0.3
python-dateutil~=2.8.2
markdownify~=0.10.3
xmltodict~=0.12.0